Change thermal modifier to float, scale down asteroids freezing effects.

+++ src/main/java/micdoodle8/mods/galacticraft/api/world/IGalacticraftWorldProvider.java
@@ -71,5 +71,5 @@ public interface IGalacticraftWorldProvider
 	 * 
 	 * @return Positive integer for hot celestial bodies, negative for cold. Zero for neutral
 	 */
-	public int getThermalLevelModifier();
+	public float getThermalLevelModifier();
 }
