+++ src/main/java/micdoodle8/mods/galacticraft/planets/mars/dimension/WorldProviderMars.java
@@ -245,7 +245,7 @@ public class WorldProviderMars extends WorldProviderSpace implements IGalacticra
 	}
 
 	@Override
-	public int getThermalLevelModifier() 
+	public float getThermalLevelModifier() 
 	{
 		return -1;
 	}
